Feat:공간 기록에서도 큐레이션에 장소 스크랩 가능하도록 구현
