+++ app/feature/search/actions/revalidateKeywordSearchPlaceData.ts
@@ -1,0 +1,7 @@
+"use server";
+
+import { revalidateTag } from "next/cache";
+
+export default async function revalidateKeywordSearchPlaceData() {
+  revalidateTag(`revalidateKeywordSearchPlaceData`);
+}
