--- app/feature/search/actions/revalidateSearchPlaceText.ts
@@ -1,7 +1,0 @@
-"use server";
-
-import { revalidatePath, revalidateTag } from "next/cache";
-
-export default async function revalidateSearchPlaceText(name: string) {
-  revalidateTag(`get${name}searchText`);
-}
