+++ app/feature/search/actions/revalidateTextSearchCurationData.ts
@@ -1,0 +1,7 @@
+"use server";
+
+import { revalidateTag } from "next/cache";
+
+export default async function revalidateTextSearchCurationData() {
+  revalidateTag(`revalidateTextSearchCurationData`);
+}
