+++ app/feature/search/queries/getTextSearchPlaceData.ts
@@ -1,12 +1,14 @@
 import { getSession } from "@common/utils/session/getSession";
 import { NextRequest, NextResponse } from "next/server";
+import { SearchPlaceResponse } from "./dto/search-type";
 
-export async function POST(request: NextRequest) {
-  const body = await request.json();
+export async function getTextSearchPlaceData(
+  name: string
+): Promise<SearchPlaceResponse> {
   const auth_info = await getSession();
   const token = auth_info?.data?.accessToken;
   const res = await fetch(
-    `${process.env.NEXT_PUBLIC_SERVER_API}/api/v1/curation/filter`,
+    `${process.env.NEXT_PUBLIC_SERVER_API}/api/v1/spaces/search?sort=recent`,
     {
       method: "POST",
       headers: {
@@ -14,13 +16,13 @@ export async function POST(request: NextRequest) {
         Authorization: `Bearer ${token}`,
       },
       body: JSON.stringify({
-        keyword1: body[0],
-        keyword2: body.length > 1 && body[1],
+        name,
       }),
       cache: "no-store",
+      next: { tags: [`getTextSearchPlaceData`] },
     }
   );
   const data = await res.json();
 
-  return NextResponse.json(data);
+  return data;
 }
